// src/lib/just-enough-javascript/verify-language-level/types.ts
type SourcePosition = {
	readonly line: number;
	readonly column: number;
};
// ...
type ParseError = {
	readonly message: string;
	readonly location: SourcePosition;
};